move langspec to static/
