+++ package.json
@@ -17,7 +17,7 @@
     "lint": "eslint examples/ src/ tests/ --ext .ts",
     "fix": "eslint --fix examples/ src/ tests/ --ext .ts",
     "docs": "typedoc",
-    "prepack": "tsc && cp src/langspec.json dist/langspec.json && cp src/lib/index.ts dist/lib/index.ts && cp src/lib/contract.ts dist/lib/contract.ts && cp src/lib/lsig.ts dist/lib/lsig.ts",
+    "prepack": "tsc && cp src/static/* dist/static/ && cp src/lib/index.ts dist/lib/index.ts && cp src/lib/contract.ts dist/lib/contract.ts && cp src/lib/lsig.ts dist/lib/lsig.ts",
     "compile-abi": "bun ./src/bin/tealscript.ts --unsafe-disable-typescript tests/contracts/abi.algo.ts tests/contracts/artifacts",
     "compile-account": "bun ./src/bin/tealscript.ts --unsafe-disable-typescript tests/contracts/account.algo.ts tests/contracts/artifacts",
     "compile-binary": "bun ./src/bin/tealscript.ts --unsafe-disable-typescript tests/contracts/binary.algo.ts tests/contracts/artifacts",
