+++ src/static/langspec.json
